♻ ⚡ move everything into structs, precompute values
This commit moves from naive representations like (Point, Point) to
things like Coord, which have extra fields for commonly used (or
complicated values): for instance, the lines of a rectangle, or the
dx/dy of a line. This helps a lot of the render math, and shows about
double speed in overall render time for the drawing used here.